Refactor utility packages and move UI components into one
