+++ desktop/package.json
@@ -49,6 +49,7 @@
     "vitest": "^0.33.0"
   },
   "dependencies": {
+    "@bowser/testing": "workspace:*",
     "@headlessui/react": "^1.7.15",
     "@radix-ui/react-alert-dialog": "^1.0.4",
     "@tailwindcss/forms": "^0.5.4",
@@ -62,7 +63,6 @@
     "@types/react-dom": "^18.2.7",
     "@types/uuid": "^9.0.2",
     "bowser-server": "workspace:*",
-    "bowser-utility-testing": "workspace:*",
     "bufferutil": "^4.0.7",
     "classnames": "^2.3.2",
     "client-only": "^0.0.1",
