+++ server/app/shows/[show_id]/rundown/[rundown_id]/RundownAssets.tsx
@@ -1,10 +1,10 @@
 "use client";
 
-import { Asset, AssetTypeSchema, Rundown } from "bowser-prisma/types";
+import { Asset, AssetTypeSchema, Rundown } from "@bowser/prisma/types";
 import React, { useEffect, useRef, useState, useTransition } from "react";
-import { AssetTypeType } from "bowser-prisma/types/inputTypeSchemas/AssetTypeSchema";
+import { AssetTypeType } from "@bowser/prisma/types/inputTypeSchemas/AssetTypeSchema";
 import classNames from "classnames";
-import Button from "@/components/ui/button";
+import Button from "@bowser/components/button";
 import {
   IoAddCircleSharp,
   IoImageSharp,
@@ -22,15 +22,15 @@ import {
   DialogContent,
   DialogHeader,
   DialogTitle,
-} from "@/components/ui/dialog";
+} from "@bowser/components/dialog";
 import {
   Popover,
   PopoverContent,
   PopoverTrigger,
-} from "@/components/ui/popover";
+} from "@bowser/components/popover";
 import Spinner from "@/app/_assets/spinner.svg";
 import Image from "next/image";
-import { Media, MediaState } from "bowser-prisma/client";
+import { Media, MediaState } from "@bowser/prisma/client";
 import {
   processAssetUpload,
   removeAsset,
