+++ server/package.json
@@ -17,11 +17,10 @@
     "@aws-sdk/client-s3": "^3.374.0",
     "@aws-sdk/s3-request-presigner": "^3.375.0",
     "@aws-sdk/signature-v4-crt": "^3.370.0",
+    "@bowser/components": "workspace:*",
+    "@bowser/prisma": "workspace:*",
+    "@bowser/testing": "workspace:*",
     "@hookform/resolvers": "^3.1.1",
-    "@radix-ui/react-dialog": "^1.0.4",
-    "@radix-ui/react-dropdown-menu": "^2.0.5",
-    "@radix-ui/react-popover": "^1.0.6",
-    "@radix-ui/react-slot": "^1.0.2",
     "@tailwindcss/forms": "^0.5.4",
     "@trpc/server": "^10.34.0",
     "@types/lodash": "^4.14.196",
@@ -32,8 +31,6 @@
     "@types/react-dom": "18.2.7",
     "@typescript-eslint/eslint-plugin": "latest",
     "@typescript-eslint/parser": "latest",
-    "bowser-prisma": "workspace:*",
-    "bowser-utility-testing": "workspace:*",
     "class-variance-authority": "^0.7.0",
     "classnames": "^2.3.2",
     "clsx": "^2.0.0",
