+++ utility/components/calendar.tsx
@@ -3,8 +3,8 @@
 import * as React from "react";
 import { DayPicker } from "react-day-picker";
 
-import { cn } from "@/lib/utils";
-import { buttonVariants } from "@/components/ui/button";
+import { cn } from "./utils";
+import { buttonVariants } from "./button";
 import { IoChevronBackSharp, IoChevronForwardSharp } from "react-icons/io5";
 
 export type CalendarProps = React.ComponentProps<typeof DayPicker>;
