+++ utility/components/dialog.tsx
@@ -3,7 +3,7 @@
 import * as React from "react";
 import * as DialogPrimitive from "@radix-ui/react-dialog";
 
-import { cn } from "@/lib/utils";
+import { cn } from "./utils";
 import { IoCloseSharp } from "react-icons/io5";
 
 const Dialog = DialogPrimitive.Root;
