+++ utility/components/package.json
@@ -1,0 +1,15 @@
+{
+  "name": "@bowser/components",
+  "private": true,
+  "dependencies": {
+    "@radix-ui/react-dialog": "^1.0.4",
+    "@radix-ui/react-dropdown-menu": "^2.0.5",
+    "@radix-ui/react-popover": "^1.0.6",
+    "@radix-ui/react-slot": "^1.0.2",
+    "class-variance-authority": "^0.7.0",
+    "clsx": "^2.0.0",
+    "react-day-picker": "^8.8.1",
+    "react-icons": "^4.10.1",
+    "tailwind-merge": "^1.14.0"
+  }
+}
